Read data from one zip file entry

Read data from one zip file entry

read_zip_entry(zip, f, zl = NULL, unserialize = FALSE)

Arguments

zip
The path to the zip file or an open connection

f
The file to read from inside the zip

zl
A zip list as produced by

unserialize
Whether to unserialize the data (from rds or qs format). Otherwise raw bytes are returned.

Value

A raw array, which may have been unserialized.

Details

When unserialize=TRUE the data must a) have been written to disk with saveRDS using gzip, bzip2 or xz compression or qs::qsave AND b) zip must have been instructed not to further compress the data (which would be pointless). This is the behaviour of write.neurons in the nat package.